@@ -1,5 +1,5 @@
-﻿
-namespace Cityvizor.Importer.Domain.Abstractions.Repositories;
+﻿namespace Cityvizor.Importer.Domain.Abstractions.Repositories;
+
 public interface IRepository<TEntity> where TEntity : class
 {
     IQueryable<TEntity> Query { get; }
