@@ -1,6 +1,7 @@
 ﻿using Cityvizor.Importer.Domain.Enums;
 
 namespace Cityvizor.Importer.Domain.Entities;
+
 public class Import
 {
     public int Id { get; set; }
@@ -13,7 +14,7 @@ public class Import
     public ImportStatus Status { get; set; }
     public string? Error { get; set; }
     public DateTime? Validity { get; set; }
-    public bool Append {  get; set; }
+    public bool Append { get; set; }
     public string? Logs { get; set; }
     public ImportFormat Format { get; set; }
     public string? ImportDir { get; set; }
