@@ -1,6 +1,7 @@
 ﻿using Cityvizor.Importer.Domain.Enums;
 
 namespace Cityvizor.Importer.Writer.Dtos;
+
 internal record ImportMetadataDto(
     ImportFormat ImportFormat,
     int ImportId);
@@ -9,5 +10,5 @@ internal record GinisImportMetadataDto(
     ImportFormat ImportFormat,
     int ImportId,
     string? AccoutingFileName,
-    string? BudgetFileName) 
+    string? BudgetFileName)
     : ImportMetadataDto(ImportFormat, ImportId);
