@@ -4,23 +4,23 @@ using System.Globalization;
 using System.Text.Json;
 
 namespace Cityvizor.Importer.Writer.Services;
+
 public class FileSystemService
 {
-    internal TMetadataDto ReadMetadateJsonFile<TMetadataDto>(string filePath)
+    internal static TMetadataDto ReadMetadateJsonFile<TMetadataDto>(string filePath)
         where TMetadataDto : ImportMetadataDto
     {
         string fileContent = File.ReadAllText(filePath);
-        TMetadataDto dto = JsonSerializer.Deserialize<TMetadataDto>(fileContent) 
+        TMetadataDto dto = JsonSerializer.Deserialize<TMetadataDto>(fileContent)
             ?? throw new JsonException($"JsonSerializer.Deserialize returned null for input {fileContent}");
         return dto;
     }
 
-    internal async Task WriteToCsvAsync<TDto>(IEnumerable<TDto> dtos, string filePath)
+    internal static async Task WriteToCsvAsync<TDto>(IEnumerable<TDto> dtos, string filePath)
     {
-        using (StreamWriter streamWriter = new(filePath))
-        using (CsvWriter csvWriter = new (streamWriter, CultureInfo.InvariantCulture)) // TODO: will existing .js importer be able to read the data written with this culture? 
-        {
-            await csvWriter.WriteRecordsAsync(dtos);
-        }
+        using StreamWriter streamWriter = new(filePath);
+        // TODO: will existing .js importer be able to read the data written with invariant culture? 
+        using CsvWriter csvWriter = new(streamWriter, CultureInfo.InvariantCulture);
+        await csvWriter.WriteRecordsAsync(dtos);
     }
 }
